:sparkles: feat : 사용자 잔액 충전 기능 구현

+++ src/main/resources/db/schema.sql
@@ -2,10 +2,10 @@ CREATE TABLE IF NOT EXISTS users
 (
     id         BIGINT AUTO_INCREMENT PRIMARY KEY,
     name       VARCHAR(50) NOT NULL COMMENT '사용자 이름',
-    balance    BIGINT       NOT NULL COMMENT '잔액',
-    version    INT          NOT NULL COMMENT '낙관적락을 위한 버전',
-    created_at DATETIME     NOT NULL COMMENT '생성 시간',
-    updated_at DATETIME     NOT NULL COMMENT '수정 시간'
+    balance    BIGINT      NOT NULL COMMENT '잔액',
+    version    BIGINT      NOT NULL COMMENT '낙관적락을 위한 버전',
+    created_at DATETIME    NOT NULL COMMENT '생성 시간',
+    updated_at DATETIME    NOT NULL COMMENT '수정 시간'
 );
 
 CREATE TABLE IF NOT EXISTS balance_histories
